requirements.txt
langchain_community==0.3.24
langchain_core==0.3.64
langchain_ollama==0.3.3
numpy==2.2.6
openai_whisper==20240930
pyaudio==0.2.14
pydantic==2.11.5
pydub==0.25.1
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
soundfile==0.13.1
speechrecognition==3.14.3
streamlit==1.45.1
torch==2.7.0
